+++ theLastHourEscapeRoom/src/modules/computer/ComputerFactory.java
@@ -28,6 +28,7 @@ import java.util.Set;
 import level.LastHourLevel;
 import modules.usbstick.UsbStickColor;
 import modules.usbstick.UsbStickItem;
+import util.LastHourQuestLogUtil;
 import util.LastHourSounds;
 import util.Lore;
 
@@ -70,7 +71,7 @@ public class ComputerFactory {
                     (eInteract, who) -> {
                       DrawComponent dc = entity.fetch(DrawComponent.class).orElseThrow();
                       if (dc.currentStateName().equals(LastHourLevel.PC_STATE_OFF)) {
-                        LastHourLevel.addRestorePowerQuestLogEntry();
+                        LastHourQuestLogUtil.addRestorePowerQuestLogEntry();
                         DialogFactory.showOkDialog(
                             "This seems to be "
                                 + Lore.ScientistNameShort
@@ -91,7 +92,7 @@ public class ComputerFactory {
                           state != null && state.state().hasReached(ComputerProgress.LOGGED_IN);
                       boolean usbAlreadyInserted = state != null && state.usbInserted();
                       if (!isLoggedIn) {
-                        LastHourLevel.addLoginNeededQuestLogEntry();
+                        LastHourQuestLogUtil.addLoginNeededQuestLogEntry();
                       }
                       if (!usbAlreadyInserted
                           && !isInfected
@@ -186,8 +187,8 @@ public class ComputerFactory {
       UsbStickItem.BaseUsbStick stick, Entity pcEntity, Entity who) {
     if (stick.color() == UsbStickColor.Blue) {
       LOGGER.info("Correct USB stick inserted: " + stick.color().displayName());
-      LastHourLevel.addUsbUsedQuestLogEntry();
-      LastHourLevel.addUsbRecoveredDataQuestLogEntry();
+      LastHourQuestLogUtil.addUsbUsedQuestLogEntry();
+      LastHourQuestLogUtil.addUsbRecoveredDataQuestLogEntry();
       // Remove the stick from inventory and mark as inserted
       who.fetch(InventoryComponent.class).ifPresent(inv -> inv.removeOne(stick));
       ComputerStateComponent.setUsbInserted(true);
@@ -195,8 +196,8 @@ public class ComputerFactory {
     } else {
       LOGGER.info(
           "Wrong USB stick inserted: " + stick.color().displayName() + " - triggering virus");
-      LastHourLevel.addUsbUsedQuestLogEntry();
-      LastHourLevel.addVirusWarningQuestLogEntry();
+      LastHourQuestLogUtil.addUsbUsedQuestLogEntry();
+      LastHourQuestLogUtil.addVirusWarningQuestLogEntry();
       ComputerStateComponent.setInfection(true);
       ComputerStateComponent.setVirusType(Lore.UnknownDeviceVirusType);
       openComputerDialog(pcEntity, who);
@@ -281,7 +282,7 @@ public class ComputerFactory {
                 }
 
                 if (newState.isInfected() && !wasInfected) {
-                  LastHourLevel.addVirusWarningQuestLogEntry();
+                  LastHourQuestLogUtil.addVirusWarningQuestLogEntry();
                   Game.add(
                       EmoteFactory.createEmote(
                           LastHourLevel.getInstance().getPoint("pc-main").translate(1f, 1.5f),
@@ -290,7 +291,7 @@ public class ComputerFactory {
                 }
                 if (!previousState.state().hasReached(ComputerProgress.LOGGED_IN)
                     && newState.state().hasReached(ComputerProgress.LOGGED_IN)) {
-                  LastHourLevel.addMailReviewQuestLogEntry();
+                  LastHourQuestLogUtil.addMailReviewQuestLogEntry();
                 }
               });
         });
